+++ src/lib/components/Avatar/Avatar.spec.tsx
@@ -6,7 +6,7 @@ import { Avatar } from './Avatar';
 
 describe('Components / Avatar', () => {
   describe('Theme', () => {
-    it('should use custom classes', () => {
+    it('should use custom sizes', () => {
       const theme: CustomFlowbiteTheme = {
         avatar: {
           size: {
@@ -22,6 +22,28 @@ describe('Components / Avatar', () => {
 
       expect(img()).toHaveClass('h-64 w-64');
     });
+
+    it('should use custom colors', () => {
+      const theme: CustomFlowbiteTheme = {
+        avatar: {
+          color: {
+            rose: 'ring-rose-500 dark:ring-rose-400',
+          },
+        },
+      };
+      render(
+        <Flowbite theme={{ theme }}>
+          <Avatar
+            bordered
+            color="rose"
+            img="https://flowbite.com/docs/images/people/profile-picture-5.jpg"
+            alt="Your avatar"
+          />
+        </Flowbite>,
+      );
+
+      expect(img()).toHaveClass('ring-rose-500 dark:ring-rose-400');
+    });
   });
   describe('Placeholder', () => {
     it('should display placeholder initials', () => {
@@ -32,6 +54,17 @@ describe('Components / Avatar', () => {
       );
 
       expect(initialsPlaceholder()).toHaveTextContent('RR');
+    });
+  });
+  describe('Image', () => {
+    it('should support custom image elements', () => {
+      render(
+        <Flowbite>
+          <Avatar img={(props) => <img referrerPolicy="no-referrer" {...props} />} />
+        </Flowbite>,
+      );
+
+      expect(img()).toHaveAttribute('referrerpolicy', 'no-referrer');
     });
   });
 });