+++ src/lib/components/Avatar/index.ts
@@ -1,1 +1,3 @@
 export * from './Avatar';
+export type { AvatarGroupProps, FlowbiteAvatarGroupTheme } from './AvatarGroup';
+export type { AvatarGroupdCounterProps, FlowbiteAvatarGroupCounterTheme } from './AvatarGroupCounter';