+++ src/lib/components/Carousel/Carousel.tsx
@@ -6,6 +6,32 @@ import ScrollContainer from 'react-indiana-drag-scroll';
 import windowExists from '../../helpers/window-exists';
 import { useTheme } from '../Flowbite/ThemeContext';
 
+export interface FlowbiteCarouselTheme {
+  base: string;
+  indicators: {
+    active: {
+      off: string;
+      on: string;
+    };
+    base: string;
+    wrapper: string;
+  };
+  item: {
+    base: string;
+    wrapper: string;
+  };
+  control: {
+    base: string;
+    icon: string;
+  };
+  leftControl: string;
+  rightControl: string;
+  scrollContainer: {
+    base: string;
+    snap: string;
+  };
+}
+
 export interface CarouselProps extends PropsWithChildren<ComponentProps<'div'>> {
   indicators?: boolean;
   leftControl?: ReactNode;
@@ -43,17 +69,18 @@ export const Carousel: FC<CarouselProps> = ({
 
   const navigateTo = useCallback(
     (item: number) => () => {
+      if (!items) return;
       item = (item + items.length) % items.length;
       if (carouselContainer.current) {
         carouselContainer.current.scrollLeft = carouselContainer.current.clientWidth * item;
       }
       setActiveItem(item);
     },
-    [items.length],
+    [items],
   );
 
   useEffect(() => {
-    if (carouselContainer.current && !isDragging) {
+    if (carouselContainer.current && !isDragging && carouselContainer.current.scrollLeft !== 0) {
       setActiveItem(Math.round(carouselContainer.current.scrollLeft / carouselContainer.current.clientWidth));
     }
   }, [isDragging]);
@@ -96,7 +123,7 @@ export const Carousel: FC<CarouselProps> = ({
       </ScrollContainer>
       {indicators && (
         <div className={theme.indicators.wrapper}>
-          {items.map(
+          {items?.map(
             (_, index): JSX.Element => (
               <button
                 key={index}
@@ -111,26 +138,31 @@ export const Carousel: FC<CarouselProps> = ({
           )}
         </div>
       )}
-      <div className={theme.leftControl}>
-        <button
-          className="group"
-          data-testid="carousel-left-control"
-          onClick={navigateTo(activeItem - 1)}
-          type="button"
-        >
-          {leftControl ? leftControl : <DefaultLeftControl />}
-        </button>
-      </div>
-      <div className={theme.rightControl}>
-        <button
-          className="group"
-          data-testid="carousel-right-control"
-          onClick={navigateTo(activeItem + 1)}
-          type="button"
-        >
-          {rightControl ? rightControl : <DefaultRightControl />}
-        </button>
-      </div>
+
+      {items && (
+        <>
+          <div className={theme.leftControl}>
+            <button
+              className="group"
+              data-testid="carousel-left-control"
+              onClick={navigateTo(activeItem - 1)}
+              type="button"
+            >
+              {leftControl ? leftControl : <DefaultLeftControl />}
+            </button>
+          </div>
+          <div className={theme.rightControl}>
+            <button
+              className="group"
+              data-testid="carousel-right-control"
+              onClick={navigateTo(activeItem + 1)}
+              type="button"
+            >
+              {rightControl ? rightControl : <DefaultRightControl />}
+            </button>
+          </div>
+        </>
+      )}
     </div>
   );
 };