+++ src/lib/components/Dropdown/Dropdown.tsx
@@ -4,34 +4,15 @@ import { HiOutlineChevronDown, HiOutlineChevronLeft, HiOutlineChevronRight, HiOu
 import { uuid } from '../../helpers/uuid';
 import type { ButtonProps } from '../Button';
 import { Button } from '../Button';
-import type { FloatingProps } from '../Floating';
+import type { FloatingProps, FlowbiteFloatingTheme } from '../Floating';
 import { Floating } from '../Floating';
 import { useTheme } from '../Flowbite/ThemeContext';
 import { DropdownDivider } from './DropdownDivider';
 import { DropdownHeader } from './DropdownHeader';
 import { DropdownItem } from './DropdownItem';
 
-export interface FlowbiteDropdownTheme {
-  target: string;
-  base: string;
-  animation: string;
-  hidden: string;
-  style: {
-    dark: string;
-    light: string;
-    auto: string;
-  };
+export interface FlowbiteDropdownFloatingTheme extends FlowbiteFloatingTheme {
   header: string;
-  content: string;
-  arrow: {
-    base: string;
-    style: {
-      dark: string;
-      light: string;
-      auto: string;
-    };
-    placement: string;
-  };
   item: {
     base: string;
     icon: string;
@@ -39,11 +20,19 @@ export interface FlowbiteDropdownTheme {
   divider: string;
 }
 
+export interface FlowbiteDropdownTheme {
+  floating: FlowbiteDropdownFloatingTheme;
+  content: string;
+  inlineWrapper: string;
+  arrowIcon: string;
+}
+
 export interface DropdownProps extends PropsWithChildren<Pick<FloatingProps, 'placement' | 'trigger'>>, ButtonProps {
   label: ReactNode;
   inline?: boolean;
   floatingArrow?: boolean;
   arrowIcon?: boolean;
+  dismissOnClick?: boolean;
 }
 
 const icons: Record<string, FC<ComponentProps<'svg'>>> = {
@@ -53,7 +42,7 @@ const icons: Record<string, FC<ComponentProps<'svg'>>> = {
   left: HiOutlineChevronLeft,
 };
 
-const DropdownComponent: FC<DropdownProps> = ({ children, className, ...props }) => {
+const DropdownComponent: FC<DropdownProps> = ({ children, className, dismissOnClick = true, ...props }) => {
   const theme = useTheme().theme.dropdown;
   const theirProps = props as DropdownProps;
   const {
@@ -80,12 +69,12 @@ const DropdownComponent: FC<DropdownProps> = ({ children, className, ...props })
       return React.cloneElement(node, {
         onClick: () => {
           node.props.onClick?.();
-          setCloseRequestKey(uuid());
+          dismissOnClick && setCloseRequestKey(uuid());
         },
       } as any);
     if (node.props.children && typeof node.props.children === 'object') {
       return React.cloneElement(node, {
-        // @ts-expect-error unknown error
+        // @ts-ignore
         children: Children.map(node.props.children, attachCloseListener),
       });
     }