+++ src/lib/components/Floating/Floating.tsx
@@ -1,6 +1,7 @@
 import type { Placement } from '@floating-ui/core';
 import {
   autoUpdate,
+  safePolygon,
   useClick,
   useFloating,
   useFocus,
@@ -12,7 +13,6 @@ import classNames from 'classnames';
 import type { ComponentProps, FC, PropsWithChildren, ReactNode } from 'react';
 import { useEffect, useRef, useState } from 'react';
 import { getArrowPlacement, getMiddleware, getPlacement } from '../../helpers/floating';
-import { FlowbiteDropdownTheme } from '../Dropdown';
 
 export interface FlowbiteFloatingTheme {
   target: string;
@@ -38,7 +38,7 @@ export interface FlowbiteFloatingTheme {
 
 export interface FloatingProps extends PropsWithChildren<Omit<ComponentProps<'div'>, 'style'>> {
   content: ReactNode;
-  theme: FlowbiteFloatingTheme | FlowbiteDropdownTheme;
+  theme: FlowbiteFloatingTheme;
   placement?: 'auto' | Placement;
   trigger?: 'hover' | 'click';
   style?: 'dark' | 'light' | 'auto';
@@ -87,7 +87,10 @@ export const Floating: FC<FloatingProps> = ({
   const { getFloatingProps, getReferenceProps } = useInteractions([
     useClick(context, { enabled: trigger === 'click' }),
     useFocus(context),
-    useHover(context, { enabled: trigger === 'hover' }),
+    useHover(context, {
+      enabled: trigger === 'hover',
+      handleClose: safePolygon(),
+    }),
     useRole(context, { role: 'tooltip' }),
   ]);